Move bench under test folder
