+++ .gitignore
@@ -2,7 +2,7 @@
 /types/
 /dist/
 /dist_type/
-/bench/results/
+/test/bench/results/
 *.es.js
 *.js.map
 node_modules
