+++ test/bench/benchmarks/filter_create.ts
@@ -1,6 +1,6 @@
 import Benchmark from '../lib/benchmark';
 
-import createFilter from '../../src/style-spec/feature_filter';
+import createFilter from '../../../src/style-spec/feature_filter';
 import filters from '../data/filters.json';
 
 export default class FilterCreate extends Benchmark {
