+++ test/bench/benchmarks/map_load.ts
