+++ test/bench/benchmarks/paint.ts
@@ -1,6 +1,6 @@
 import Benchmark from '../lib/benchmark';
 import createMap from '../lib/create_map';
-import type Map from '../../src/ui/map';
+import type Map from '../../../src/ui/map';
 
 const width = 1024;
 const height = 768;
