+++ test/bench/benchmarks/placement.ts
@@ -1,12 +1,11 @@
 import Benchmark from '../lib/benchmark';
 import createMap from '../lib/create_map';
-import type Map from '../../src/ui/map';
-import type {PointLike} from '../../src/ui/camera';
+import type Map from '../../../src/ui/map';
 
 const width = 1024;
 const height = 768;
 
-export default class QueryBox extends Benchmark {
+export default class Paint extends Benchmark {
     style: string;
     locations: Array<any>;
     maps: Array<Map>;
@@ -24,7 +23,8 @@ export default class QueryBox extends Benchmark {
                 width,
                 height,
                 center: location.center,
-                style: this.style
+                style: this.style,
+                idle: true
             });
         }))
             .then(maps => {
@@ -37,7 +37,17 @@ export default class QueryBox extends Benchmark {
 
     bench() {
         for (const map of this.maps) {
-            map.queryRenderedFeatures({} as PointLike);
+            const showCollisionBoxes = false;
+            const fadeDuration = 300;
+            const crossSourceCollisions = true;
+            const forceFullPlacement = true;
+
+            map.style._updatePlacement(
+                map.transform,
+                showCollisionBoxes,
+                fadeDuration,
+                crossSourceCollisions,
+                forceFullPlacement);
         }
     }
