+++ test/bench/benchmarks/query_box.ts
@@ -1,11 +1,12 @@
 import Benchmark from '../lib/benchmark';
 import createMap from '../lib/create_map';
-import type Map from '../../src/ui/map';
+import type Map from '../../../src/ui/map';
+import type {PointLike} from '../../../src/ui/camera';
 
 const width = 1024;
 const height = 768;
 
-export default class Paint extends Benchmark {
+export default class QueryBox extends Benchmark {
     style: string;
     locations: Array<any>;
     maps: Array<Map>;
@@ -23,8 +24,7 @@ export default class Paint extends Benchmark {
                 width,
                 height,
                 center: location.center,
-                style: this.style,
-                idle: true
+                style: this.style
             });
         }))
             .then(maps => {
@@ -37,17 +37,7 @@ export default class Paint extends Benchmark {
 
     bench() {
         for (const map of this.maps) {
-            const showCollisionBoxes = false;
-            const fadeDuration = 300;
-            const crossSourceCollisions = true;
-            const forceFullPlacement = true;
-
-            map.style._updatePlacement(
-                map.transform,
-                showCollisionBoxes,
-                fadeDuration,
-                crossSourceCollisions,
-                forceFullPlacement);
+            map.queryRenderedFeatures({} as PointLike);
         }
     }
