+++ test/bench/benchmarks/query_point.ts
@@ -1,7 +1,7 @@
 import Benchmark from '../lib/benchmark';
 import createMap from '../lib/create_map';
-import type Map from '../../src/ui/map';
-import type {LngLatLike} from '../../src/geo/lng_lat';
+import type Map from '../../../src/ui/map';
+import type {LngLatLike} from '../../../src/geo/lng_lat';
 
 const width = 1024;
 const height = 768;
