+++ test/bench/benchmarks/remove_paint_state.ts
