+++ test/bench/benchmarks/style_layer_create.ts
@@ -1,7 +1,7 @@
-import type {StyleSpecification} from '../../src/style-spec/types';
+import type {StyleSpecification} from '../../../src/style-spec/types';
 import Benchmark from '../lib/benchmark';
-import createStyleLayer from '../../src/style/create_style_layer';
-import deref from '../../src/style-spec/deref';
+import createStyleLayer from '../../../src/style/create_style_layer';
+import deref from '../../../src/style-spec/deref';
 import fetchStyle from '../lib/fetch_style';
 
 export default class StyleLayerCreate extends Benchmark {
