+++ test/bench/benchmarks/style_validate.ts
@@ -1,6 +1,6 @@
-import type {StyleSpecification} from '../../src/style-spec/types';
+import type {StyleSpecification} from '../../../src/style-spec/types';
 import Benchmark from '../lib/benchmark';
-import validateStyle from '../../src/style-spec/validate_style.min';
+import validateStyle from '../../../src/style-spec/validate_style.min';
 import fetchStyle from '../lib/fetch_style';
 
 export default class StyleValidate extends Benchmark {
