+++ test/bench/benchmarks/symbol_layout.ts
@@ -1,7 +1,7 @@
 import Layout from './layout';
-import SymbolBucket from '../../src/data/bucket/symbol_bucket';
-import {performSymbolLayout} from '../../src/symbol/symbol_layout';
-import {OverscaledTileID} from '../../src/source/tile_id';
+import SymbolBucket from '../../../src/data/bucket/symbol_bucket';
+import {performSymbolLayout} from '../../../src/symbol/symbol_layout';
+import {OverscaledTileID} from '../../../src/source/tile_id';
 
 export default class SymbolLayout extends Layout {
     parsedTiles: Array<any>;
