+++ test/bench/benchmarks/worker_transfer.ts
@@ -1,9 +1,9 @@
-import type {StyleSpecification} from '../../src/style-spec/types';
+import type {StyleSpecification} from '../../../src/style-spec/types';
 import Benchmark from '../lib/benchmark';
 import fetchStyle from '../lib/fetch_style';
 import TileParser from '../lib/tile_parser';
-import {OverscaledTileID} from '../../src/source/tile_id';
-import {serialize, deserialize} from '../../src/util/web_worker_transfer';
+import {OverscaledTileID} from '../../../src/source/tile_id';
+import {serialize, deserialize} from '../../../src/util/web_worker_transfer';
 
 export default class WorkerTransfer extends Benchmark {
     parser: TileParser;
