+++ test/bench/benchmarks_view.tsx
