+++ test/bench/components/StatisticsPlot.tsx
