+++ test/bench/gl-stats.ts
@@ -5,7 +5,7 @@ import {execSync} from 'child_process';
 
 const maplibreGLJSSrc = fs.readFileSync('dist/maplibre-gl.js', 'utf8');
 const maplibreGLCSSSrc = fs.readFileSync('dist/maplibre-gl.css', 'utf8');
-const benchSrc = fs.readFileSync('bench/gl-stats.html', 'utf8');
+const benchSrc = fs.readFileSync('test/bench/gl-stats.html', 'utf8');
 
 const benchHTML = benchSrc
     .replace(/<script src="..\/dist\/maplibre-gl.js"><\/script>/, `<script>${maplibreGLJSSrc}</script>`);
