+++ test/bench/lib/fetch_style.ts
@@ -1,4 +1,4 @@
-import type {StyleSpecification} from '../../src/style-spec/types';
+import type {StyleSpecification} from '../../../src/style-spec/types';
 
 export default function fetchStyle(value: string | StyleSpecification): Promise<StyleSpecification> {
     return typeof value === 'string' ?
