+++ test/bench/lib/statistics.ts
