+++ test/bench/lib/tile_parser.ts
@@ -2,18 +2,18 @@ import Protobuf from 'pbf';
 import VT from '@mapbox/vector-tile';
 import assert from 'assert';
 
-import deref from '../../src/style-spec/deref';
-import Style from '../../src/style/style';
-import {Evented} from '../../src/util/evented';
-import {RequestManager} from '../../src/util/request_manager';
-import WorkerTile from '../../src/source/worker_tile';
-import StyleLayerIndex from '../../src/style/style_layer_index';
+import deref from '../../../src/style-spec/deref';
+import Style from '../../../src/style/style';
+import {Evented} from '../../../src/util/evented';
+import {RequestManager} from '../../../src/util/request_manager';
+import WorkerTile from '../../../src/source/worker_tile';
+import StyleLayerIndex from '../../../src/style/style_layer_index';
 
-import type {StyleSpecification} from '../../src/style-spec/types';
-import type {WorkerTileResult} from '../../src/source/worker_source';
-import type {OverscaledTileID} from '../../src/source/tile_id';
-import type {TileJSON} from '../../src/types/tilejson';
-import type Map from '../../src/ui/map';
+import type {StyleSpecification} from '../../../src/style-spec/types';
+import type {WorkerTileResult} from '../../../src/source/worker_source';
+import type {OverscaledTileID} from '../../../src/source/tile_id';
+import type {TileJSON} from '../../../src/types/tilejson';
+import type Map from '../../../src/ui/map';
 
 class StubMap extends Evented {
     _requestManager: RequestManager;
