+++ test/bench/styles/index.html
@@ -12,8 +12,8 @@
 
 <body>
     <div id="benchmarks"></div>
-    <script src="/bench/styles/benchmarks_generated.js"></script>
-    <script src="/bench/benchmarks_view_generated.js"></script>
+    <script src="test/bench/styles/benchmarks_generated.js"></script>
+    <script src="test/bench/benchmarks_view_generated.js"></script>
     <script>
         window.Benchmarks.run(benchmarks);
     </script>
