+++ tsconfig.json
@@ -29,15 +29,15 @@
   "include": [
     "rollup.config.**",
     "src/**/*.ts",
-    "bench/**/*.ts",
-    "bench/**/*.tsx",
+    "test/bench/**/*.ts",
+    "test/bench/**/*.tsx",
     "build/**/*.ts"
   ],
   "exclude": [
     "node_modules",
     "lib",
     "dist",
-    "bench/**/*generated*",
+    "test/bench/**/*generated*",
     "src/**/*.test.ts"
   ]
 }
